move comp into tasks folder
